Place out request clients in their own modules.
Makes it easier to analyze which components can be generalized.
